unsigned long longs to uint64 (cosmetic)

+++ computation_struct.h
@@ -22,10 +22,10 @@ public:
 	// Hi-res gizmos
 	bool isHires; // True only for the hi-res computation
 	int mapIndex; // Index of the map to be computed with hi-res
-	unsigned long long variationsInCurrentExecute; // Current variations count, can be less at the end of hi-res computation
-	unsigned long long startVariationInCurrentExecute; // Index of the first variation in the execute
+	uint64_t variationsInCurrentExecute; // Current variations count, can be less at the end of hi-res computation
+	uint64_t startVariationInCurrentExecute; // Index of the first variation in the execute
 	int buffersPerVariation = 1; // How many times is one variation repeated (steps * buffersPerVariation equals total steps per variation)
-	unsigned long long variationsFinished; // How many variations of marshal.totalVariations have finished hi-res-computing
+	uint64_t variationsFinished; // How many variations of marshal.totalVariations have finished hi-res-computing
 
 	void Clear()
 	{
